dsl: split 2199-line lower/tests.rs into topical submodules
Break the monolithic tests file into a tests/ directory with focused
submodules (primitives, branch, materials, layout, lod, lights, deform,
control_flow, extrude_sweep_loft, parametric_surfaces). Shared helpers
(lower_src, find_mesh_node, mesh_aabb) live in tests/mod.rs. All 324
tests still pass.
