--- crates/mogen-dsl/src/lower/tests.rs
@@ -1,2199 +1,0 @@
-use super::*;
-use crate::parser::parse;
-use glam::Vec3;
-
-fn lower_src(src: &str) -> SceneGraph {
-    let ast = parse(src).expect("parse");
-    lower(&ast).expect("lower")
-}
-
-fn find_mesh_node<'a>(g: &'a SceneGraph, name: &str) -> &'a mogen_core::SceneNode {
-    g.nodes
-        .iter()
-        .find(|n| n.name == name)
-        .unwrap_or_else(|| panic!("no node named {name}"))
-}
-
-#[test]
-fn lowers_every_new_primitive() {
-    // One scene that exercises every new primitive kind end-to-end:
-    // parse → validate attrs → lower → mesh attached to node.
-    let g = lower_src(
-        r#"
-        scene {
-          wedge         "w" (size=[1, 0.5, 1])
-          frustum       "f" (bottom=[1, 1], top=[0.5, 0.5], height=1)
-          tube          "t" (outer=0.5, inner=0.3, height=1)
-          hemisphere    "h" (radius=0.5)
-          half_cylinder "hc" (radius=0.5, height=1)
-          torus_arc     "ta" (major=0.5, minor=0.1, arc=90)
-          ellipsoid     "e" (size=[1, 0.5, 0.8])
-        }
-    "#,
-    );
-    for name in ["w", "f", "t", "h", "hc", "ta", "e"] {
-        let n = find_mesh_node(&g, name);
-        assert!(n.mesh.is_some(), "{name} has no mesh");
-        let mesh = n.mesh.as_ref().unwrap();
-        assert!(!mesh.positions.is_empty(), "{name} mesh has no positions");
-        assert!(!mesh.indices.is_empty(), "{name} mesh has no indices");
-        // Default connectors were populated.
-        assert!(!n.connectors.is_empty(), "{name} has no default connectors");
-    }
-}
-
-#[test]
-fn tube_has_inner_and_outer_walls() {
-    let g = lower_src(
-        r#"scene { tube "t" (outer=1.0, inner=0.5, height=1.0) }"#,
-    );
-    let n = find_mesh_node(&g, "t");
-    let mesh = n.mesh.as_ref().unwrap();
-    // Some verts at outer radius, some at inner radius — cheap "is hollow" check.
-    let has_outer = mesh.positions.iter().any(|p| (p[0] * p[0] + p[2] * p[2]).sqrt() > 0.9);
-    let has_inner = mesh.positions.iter().any(|p| {
-        let r = (p[0] * p[0] + p[2] * p[2]).sqrt();
-        r > 0.4 && r < 0.6
-    });
-    assert!(has_outer, "tube is missing outer wall");
-    assert!(has_inner, "tube is missing inner wall");
-}
-
-#[test]
-fn hemisphere_has_base_at_origin() {
-    let g = lower_src(r#"scene { hemisphere "h" (radius=1.0) }"#);
-    let mesh = find_mesh_node(&g, "h").mesh.as_ref().unwrap();
-    // Base cap sits on y=0; apex at y=+radius.
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    assert!((min_y).abs() < 1e-5, "expected base at y=0, got {min_y}");
-    assert!((max_y - 1.0).abs() < 1e-5, "expected apex at y=radius, got {max_y}");
-}
-
-#[test]
-fn wedge_slope_connector_faces_up_and_forward() {
-    let g = lower_src(r#"scene { wedge "w" (size=[1.0, 1.0, 1.0]) }"#);
-    let n = find_mesh_node(&g, "w");
-    let slope = n
-        .connectors
-        .iter()
-        .find(|c| c.name == "slope")
-        .expect("wedge missing slope connector");
-    // Connector rotation turns +Y into the connector's outward dir.
-    let dir = slope.rotation * Vec3::Y;
-    assert!(dir.y > 0.0 && dir.z > 0.0, "slope normal should point +Y and +Z, got {dir:?}");
-}
-
-#[test]
-fn leaf_card_emits_cross_quads_with_bottom_at_origin() {
-    let g = lower_src(
-        r#"scene { leaf_card "l" (size=[0.4, 0.5], cards=2) }"#,
-    );
-    let mesh = find_mesh_node(&g, "l").mesh.as_ref().unwrap();
-    // Two cards × 4 verts each = 8 verts. Single winding: two tris per
-    // card → 2 cards × 6 = 12 indices. The material's `double_sided`
-    // flag is what makes the leaf visible from both sides.
-    assert_eq!(mesh.positions.len(), 8);
-    assert_eq!(mesh.indices.len(), 12);
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    assert!(min_y.abs() < 1e-5, "bottom edge should sit at y=0, got {min_y}");
-    assert!((max_y - 0.5).abs() < 1e-5, "top edge should be at y=h, got {max_y}");
-    // Default connectors include `stem` at origin and `tip` at the top.
-    let n = find_mesh_node(&g, "l");
-    let names: Vec<_> = n.connectors.iter().map(|c| c.name.as_str()).collect();
-    assert!(names.contains(&"stem"), "missing stem: {names:?}");
-    assert!(names.contains(&"tip"), "missing tip: {names:?}");
-}
-
-#[test]
-fn leaf_card_with_alpha_mask_material_passes_validation() {
-    // Full pipeline: alpha-cutout leaf with double_sided rendering — the
-    // canonical recipe for game foliage.
-    let g = lower_src(
-        r#"
-        material "leaf" (
-            color=[0.3, 0.6, 0.2],
-            alpha_mode="mask",
-            alpha_cutoff=0.5,
-            double_sided=1
-        )
-        scene {
-          leaf_card "l" (size=[0.4, 0.4], cards=3, mat="leaf")
-        }
-        "#,
-    );
-    let leaf = find_mesh_node(&g, "l");
-    assert!(leaf.mesh.is_some());
-    let mid = leaf.material.expect("leaf material bound");
-    let m = &g.materials[mid.0 as usize];
-    assert_eq!(m.alpha_mode, mogen_core::AlphaMode::Mask);
-    assert!(m.double_sided);
-}
-
-#[test]
-fn skeleton_nested_inside_group_is_lowered() {
-    // Same shape as the businessman example: wrapping `use "humanoid_full" ()`
-    // in `group "humanoid" { ... }` puts the module-declared `skeleton "rig"`
-    // inside that group. The skeleton must still register so a sibling mesh
-    // with `skin="rig"` can bind to it.
-    let g = lower_src(
-        r#"
-        scene {
-          group "wrapper" {
-            skeleton "rig" {
-              bone "root" (envelope=0.5)
-            }
-            box "thing" (size=[1, 1, 1], skin="rig", bind="root")
-          }
-        }
-        "#,
-    );
-    let skin = g.find_skin("rig").expect("rig skin registered from inside group");
-    let thing = find_mesh_node(&g, "thing");
-    assert_eq!(thing.skin, Some(skin), "mesh inside the same group should bind to rig");
-}
-
-#[test]
-fn materials_nested_inside_groups_are_discovered() {
-    // A `material` decl that lands inside a wrapping `group` (the shape that
-    // module expansion produces when the user writes
-    // `group "g" { use "m" () }`) must still be findable for `mat=` lookup.
-    // Regression: collect_materials used to walk only top-level / scene-level.
-    let g = lower_src(
-        r#"
-        scene {
-          group "wrapper" {
-            material "boot" (color=[0.04, 0.04, 0.04])
-            box "shoe" (size=[1, 0.5, 1.5], mat="boot")
-          }
-        }
-        "#,
-    );
-    let mid = g.find_material("boot").expect("boot material registered");
-    let shoe = find_mesh_node(&g, "shoe");
-    assert_eq!(shoe.material, Some(mid), "shoe should resolve mat=\"boot\"");
-}
-
-#[test]
-fn branch_expands_to_seg_and_leaf_nodes() {
-    // depth=2, splits=2 → 1 + 2 + 4 = 7 segments and 4 leaves.
-    let g = lower_src(
-        r#"
-        material "bark" (color=[0.4, 0.25, 0.15])
-        material "leaf" (color=[0.3, 0.6, 0.2], alpha_mode="mask", double_sided=1)
-        scene {
-          branch "tree" (
-            length=1.0, radius=0.1, depth=2, splits=2,
-            length_falloff=0.7, radius_falloff=0.6,
-            branch_angle=30, jitter=0.0,
-            leaves=1, leaf_size=0.3, leaf_mat="leaf",
-            mat="bark"
-          )
-        }
-        "#,
-    );
-    let segs = g.nodes.iter().filter(|n| n.kind == "branch_seg").count();
-    let leaves = g.nodes.iter().filter(|n| n.kind == "leaf_card").count();
-    assert_eq!(segs, 7, "expected 7 branch segments at depth=2 splits=2, got {segs}");
-    assert_eq!(leaves, 4, "expected 4 leaf cards at depth=0 tips, got {leaves}");
-    // Bark inherited on segments; explicit leaf material on leaves.
-    let bark = g.find_material("bark").expect("bark");
-    let leaf = g.find_material("leaf").expect("leaf");
-    for n in &g.nodes {
-        if n.kind == "branch_seg" {
-            assert_eq!(n.material, Some(bark), "segment {} should inherit bark", n.name);
-        } else if n.kind == "leaf_card" {
-            assert_eq!(n.material, Some(leaf), "leaf {} should bind leaf material", n.name);
-        }
-    }
-}
-
-#[test]
-fn branch_is_deterministic_for_a_given_seed() {
-    let a = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=3, splits=3, jitter=0.5, seed=7) }"#,
-    );
-    let b = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=3, splits=3, jitter=0.5, seed=7) }"#,
-    );
-    assert_eq!(a.nodes.len(), b.nodes.len());
-    for (na, nb) in a.nodes.iter().zip(b.nodes.iter()) {
-        assert_eq!(na.kind, nb.kind, "kind diverges for seeded branch");
-        assert_eq!(na.transform.translation, nb.transform.translation);
-    }
-}
-
-#[test]
-fn branch_seed_changes_geometry() {
-    let a = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=3, splits=3, jitter=0.5, seed=1) }"#,
-    );
-    let b = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=3, splits=3, jitter=0.5, seed=2) }"#,
-    );
-    // Same node count, but at least one transform should differ.
-    assert_eq!(a.nodes.len(), b.nodes.len());
-    let any_diff = a
-        .nodes
-        .iter()
-        .zip(b.nodes.iter())
-        .any(|(x, y)| x.transform.translation != y.transform.translation);
-    assert!(any_diff, "different seeds should produce different forks");
-}
-
-#[test]
-fn branch_no_leaves_when_disabled() {
-    let g = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=2, splits=2, leaves=0) }"#,
-    );
-    let leaves = g.nodes.iter().filter(|n| n.kind == "leaf_card").count();
-    assert_eq!(leaves, 0, "leaves=0 should suppress leaf cards");
-}
-
-#[test]
-fn branch_leaves_align_to_branch_tip() {
-    // With branch_angle=0 and bend=0 every segment grows along +Y, so a
-    // depth=1 split should keep its single leaf pointing up — the leaf
-    // card's local +Y should resolve to world +Y after composition.
-    let g = lower_src(
-        r#"scene {
-            branch "t" (
-                length=1, radius=0.1, depth=1, splits=1,
-                branch_angle=0, bend=0, tropism=0, jitter=0,
-                leaves=1, leaf_size=0.2
-            )
-        }"#,
-    );
-    // Find the leaf node and walk its world transform.
-    let world = g.world_transforms();
-    let leaf_idx = g
-        .nodes
-        .iter()
-        .position(|n| n.kind == "leaf_card")
-        .expect("leaf card present");
-    let m = world[leaf_idx];
-    let up = m.transform_vector3(Vec3::Y).normalize();
-    assert!(up.y > 0.95, "leaf +Y should align with world +Y, got {up:?}");
-}
-
-#[test]
-fn branch_segments_are_marked_non_editable() {
-    let g = lower_src(
-        r#"scene { branch "t" (length=1, radius=0.1, depth=1, splits=2) }"#,
-    );
-    for n in &g.nodes {
-        if matches!(n.kind.as_str(), "branch_seg" | "leaf_card") {
-            assert!(!n.editable, "{} should be non-editable", n.name);
-        }
-    }
-    // Wrapper itself stays editable so the user can tweak `branch` attrs.
-    let wrapper = g
-        .nodes
-        .iter()
-        .find(|n| n.kind == "branch")
-        .expect("wrapper present");
-    assert!(wrapper.editable, "branch wrapper should remain editable");
-}
-
-#[test]
-fn lowers_every_organic_primitive() {
-    // End-to-end check of the four organic-shape primitives. Uses nested
-    // list literals (`[[x,y,z], ...]`, `[[r,y], ...]`) to confirm the
-    // grammar extension landed.
-    let g = lower_src(
-        r#"
-        scene {
-          superellipsoid "se"   (size=[1, 0.8, 1], ew=0.5, ns=1)
-          curved_plane   "leaf" (size=[0.4, 1.0], bend_u=20, bend_v=40)
-          lathe          "vase" (profile=[[0.0, -0.5], [0.4, -0.3], [0.5, 0.0], [0.3, 0.4], [0.0, 0.5]])
-          spline_tube    "ban"  (points=[[0, 0, 0], [0.3, 0.2, 0], [0.5, 0.1, 0], [0.6, -0.1, 0]],
-                                 radii=[0.08, 0.12, 0.10, 0.05])
-        }
-    "#,
-    );
-    for name in ["se", "leaf", "vase", "ban"] {
-        let n = find_mesh_node(&g, name);
-        assert!(n.mesh.is_some(), "{name} has no mesh");
-        let mesh = n.mesh.as_ref().unwrap();
-        assert!(!mesh.positions.is_empty(), "{name} mesh has no positions");
-        assert!(!mesh.indices.is_empty(), "{name} mesh has no indices");
-        assert_eq!(mesh.positions.len(), mesh.normals.len(), "{name} normals arity mismatch");
-    }
-}
-
-#[test]
-fn superellipsoid_boxy_exponent_fills_corners() {
-    // ew, ns > 1 push the shape toward a box — corner vertices sit close to
-    // the declared size bounds, unlike a sphere which tucks them inward.
-    let g = lower_src(
-        r#"scene { superellipsoid "s" (size=[1.0, 1.0, 1.0], ew=3.0, ns=3.0, rings=24, segments=32) }"#,
-    );
-    let mesh = find_mesh_node(&g, "s").mesh.as_ref().unwrap();
-    // Find the vertex nearest the +X+Y+Z corner and check it's close to [0.5, 0.5, 0.5].
-    let max_corner = mesh
-        .positions
-        .iter()
-        .map(|p| (p[0] + p[1] + p[2], *p))
-        .fold((f32::NEG_INFINITY, [0.0; 3]), |acc, x| if x.0 > acc.0 { x } else { acc })
-        .1;
-    // Sphere would give ~0.29 on each axis; boxy should be > 0.4.
-    assert!(max_corner[0] > 0.4 && max_corner[1] > 0.4 && max_corner[2] > 0.4,
-        "boxy superellipsoid should reach corners, got {max_corner:?}");
-}
-
-#[test]
-fn superellipsoid_faces_wind_outward() {
-    // Face winding must match vertex normals so back-face culling shows the
-    // outside. Regression: ring 0 sits at the south pole, so the sphere's
-    // north-first winding would flip every triangle.
-    let g = lower_src(
-        r#"scene { superellipsoid "s" (size=[1.0, 1.0, 1.0], ew=1.0, ns=1.0, rings=12, segments=16) }"#,
-    );
-    let mesh = find_mesh_node(&g, "s").mesh.as_ref().unwrap();
-    let mut checked = 0usize;
-    let mut aligned = 0usize;
-    for tri in mesh.indices.chunks_exact(3) {
-        let (ia, ib, ic) = (tri[0] as usize, tri[1] as usize, tri[2] as usize);
-        let a = Vec3::from(mesh.positions[ia]);
-        let b = Vec3::from(mesh.positions[ib]);
-        let c = Vec3::from(mesh.positions[ic]);
-        let face = (b - a).cross(c - a);
-        // Polar caps collapse to a point — skip those degenerate tris.
-        if face.length_squared() < 1e-12 {
-            continue;
-        }
-        checked += 1;
-        let avg = (Vec3::from(mesh.normals[ia])
-            + Vec3::from(mesh.normals[ib])
-            + Vec3::from(mesh.normals[ic]))
-            / 3.0;
-        if face.dot(avg) > 0.0 {
-            aligned += 1;
-        }
-    }
-    assert!(checked > 0, "expected non-degenerate superellipsoid faces");
-    assert_eq!(aligned, checked, "all non-degenerate superellipsoid faces should wind outward");
-}
-
-#[test]
-fn curved_plane_bends_toward_positive_y() {
-    // Positive bend_u lifts the left/right edges. The centre stays near y=0;
-    // the edges sit well above y=0.
-    let g = lower_src(
-        r#"scene { curved_plane "l" (size=[1.0, 0.2], bend_u=90, segments_u=16) }"#,
-    );
-    let mesh = find_mesh_node(&g, "l").mesh.as_ref().unwrap();
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    assert!(max_y > 0.05, "bent plane should lift edges above y=0, got max_y={max_y}");
-    assert!(min_y.abs() < 1e-4, "unbent center should still sit at y=0, got min_y={min_y}");
-}
-
-#[test]
-fn lathe_revolves_around_y() {
-    // A flat profile `[0.5, 0.0]` for two rows makes a closed cylinder;
-    // every vertex on the side wall lands at radius ≈ 0.5.
-    let g = lower_src(
-        r#"scene { lathe "l" (profile=[[0.5, -0.5], [0.5, 0.5]], segments=16) }"#,
-    );
-    let mesh = find_mesh_node(&g, "l").mesh.as_ref().unwrap();
-    let side_verts: Vec<_> = mesh
-        .positions
-        .iter()
-        .filter(|p| (p[0] * p[0] + p[2] * p[2]).sqrt() > 0.4)
-        .collect();
-    assert!(!side_verts.is_empty(), "lathe should have side-wall verts");
-    for p in side_verts {
-        let r = (p[0] * p[0] + p[2] * p[2]).sqrt();
-        assert!((r - 0.5).abs() < 1e-4, "side-wall radius should be 0.5, got {r}");
-    }
-}
-
-#[test]
-fn spline_tube_follows_control_points() {
-    // Straight tube along Y should yield every vertex in a narrow X-band
-    // around the axis.
-    let g = lower_src(
-        r#"scene { spline_tube "t" (points=[[0,0,0],[0,0.5,0],[0,1,0]], radius=0.1, segments=8, samples=4) }"#,
-    );
-    let mesh = find_mesh_node(&g, "t").mesh.as_ref().unwrap();
-    for p in &mesh.positions {
-        let r = (p[0] * p[0] + p[2] * p[2]).sqrt();
-        assert!(r < 0.12, "straight tube along Y should stay near the axis, got r={r}");
-    }
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    assert!(min_y < 0.05 && max_y > 0.95, "tube should span y∈[0, 1], got [{min_y}, {max_y}]");
-}
-
-#[test]
-fn spline_tube_exposes_start_and_end_connectors() {
-    let g = lower_src(
-        r#"scene { spline_tube "t" (points=[[0,0,0],[0.5,0.5,0],[1,0,0]], radius=0.05) }"#,
-    );
-    let n = find_mesh_node(&g, "t");
-    let names: Vec<_> = n.connectors.iter().map(|c| c.name.as_str()).collect();
-    assert!(names.contains(&"start"), "connectors: {names:?}");
-    assert!(names.contains(&"end"), "connectors: {names:?}");
-}
-
-#[test]
-fn csg_inherits_first_operand_material_when_unset() {
-    let g = lower_src(
-        r#"
-        material "brick" (color=[0.7, 0.3, 0.2])
-        material "soot"  (color=[0.1, 0.1, 0.1])
-        scene {
-          difference "dome" {
-            hemisphere "outer" (radius=0.6, mat="brick")
-            hemisphere "inner" (radius=0.5, mat="soot")
-          }
-        }
-        "#,
-    );
-    let dome = find_mesh_node(&g, "dome");
-    let brick = g.find_material("brick").expect("brick material");
-    assert_eq!(dome.material, Some(brick),
-        "CSG should inherit first operand's material when own mat is absent");
-}
-
-#[test]
-fn csg_own_material_wins_over_operand() {
-    let g = lower_src(
-        r#"
-        material "brick" (color=[0.7, 0.3, 0.2])
-        material "stone" (color=[0.5, 0.5, 0.5])
-        scene {
-          difference "dome" (mat="stone") {
-            hemisphere "outer" (radius=0.6, mat="brick")
-            hemisphere "inner" (radius=0.5)
-          }
-        }
-        "#,
-    );
-    let dome = find_mesh_node(&g, "dome");
-    let stone = g.find_material("stone").expect("stone material");
-    assert_eq!(dome.material, Some(stone),
-        "explicit mat on CSG node must win over first-operand inheritance");
-}
-
-#[test]
-fn material_inherits_from_parent_to_unmarked_children() {
-    // Regression: setting `mat` on a `solid` (or any group) must apply to
-    // children that don't override it. Without this the merge pass groups
-    // them under None and the merged leaf renders untextured.
-    let g = lower_src(
-        r#"
-        material "wood" (color=[0.45, 0.28, 0.15])
-        scene {
-          solid "crate" (mat="wood") {
-            box "floor" (size=1)
-            group "walls" {
-              box "left"  (pos=[-0.5, 0, 0], size=[0.1, 1, 1])
-              box "right" (pos=[ 0.5, 0, 0], size=[0.1, 1, 1])
-            }
-          }
-        }
-        "#,
-    );
-    let wood = g.find_material("wood").expect("wood material");
-    for name in ["crate", "floor", "walls", "left", "right"] {
-        assert_eq!(
-            find_mesh_node(&g, name).material,
-            Some(wood),
-            "{name} should inherit wood from parent solid"
-        );
-    }
-}
-
-#[test]
-fn inherited_material_drives_primitive_uv_mode() {
-    // Regression: uv_mode is baked into the mesh at primitive generation.
-    // When a child inherits its material from a parent, the primitive must
-    // see the inherited material's uv_mode — otherwise size=2 boxes render
-    // with tile-mode UVs [0, 2] when the author asked for fit-mode [0, 1].
-    let g = lower_src(
-        r#"
-        material "sign" (color=[1, 1, 1], uv_mode="fit")
-        scene {
-          group "billboard" (mat="sign") {
-            box "face" (size=2)
-          }
-        }
-        "#,
-    );
-    let mesh = find_mesh_node(&g, "face").mesh.as_ref().unwrap();
-    let max_uv = mesh.uvs.iter().flat_map(|u| u.iter().copied()).fold(f32::NEG_INFINITY, f32::max);
-    assert!(
-        max_uv <= 1.0 + 1e-5,
-        "expected fit-mode UVs capped at 1.0, got max={max_uv} — uv_mode didn't inherit"
-    );
-}
-
-#[test]
-fn child_material_overrides_inherited_parent_material() {
-    let g = lower_src(
-        r#"
-        material "wood"  (color=[0.45, 0.28, 0.15])
-        material "metal" (color=[0.7, 0.7, 0.7])
-        scene {
-          group "g" (mat="wood") {
-            box "a" (size=1)
-            box "b" (size=1, mat="metal")
-          }
-        }
-        "#,
-    );
-    let wood = g.find_material("wood").expect("wood material");
-    let metal = g.find_material("metal").expect("metal material");
-    assert_eq!(find_mesh_node(&g, "a").material, Some(wood));
-    assert_eq!(find_mesh_node(&g, "b").material, Some(metal));
-}
-
-#[test]
-fn solid_lowers_as_tagged_group() {
-    let g = lower_src(
-        r#"
-        material "stone" (color=[0.8, 0.8, 0.8])
-        scene {
-          solid "shell" (mat="stone") {
-            box "a" (size=1)
-            box "b" (pos=[0.5, 0, 0], size=1)
-          }
-        }
-        "#,
-    );
-    let shell = find_mesh_node(&g, "shell");
-    assert!(shell.mesh.is_none(), "solid itself has no mesh");
-    assert!(shell.tags.iter().any(|t| t == "solid"));
-    assert!(!shell.tags.iter().any(|t| t == "cleanup=coplanar"));
-    assert_eq!(shell.children.len(), 2, "children are preserved");
-}
-
-#[test]
-fn solid_records_cleanup_coplanar_tag() {
-    let g = lower_src(
-        r#"
-        material "stone" (color=[0.8, 0.8, 0.8])
-        scene {
-          solid "shell" (mat="stone", cleanup="coplanar") {
-            box "a" (size=1)
-          }
-        }
-        "#,
-    );
-    let shell = find_mesh_node(&g, "shell");
-    assert!(shell.tags.iter().any(|t| t == "solid"));
-    assert!(shell.tags.iter().any(|t| t == "cleanup=coplanar"));
-}
-
-fn mesh_aabb(g: &SceneGraph, name: &str) -> (Vec3, Vec3) {
-    let mesh = find_mesh_node(g, name).mesh.as_ref().unwrap();
-    let min = mesh.positions.iter().fold(Vec3::splat(f32::INFINITY), |a, p| {
-        a.min(Vec3::from_array(*p))
-    });
-    let max = mesh.positions.iter().fold(Vec3::splat(f32::NEG_INFINITY), |a, p| {
-        a.max(Vec3::from_array(*p))
-    });
-    (min, max)
-}
-
-#[test]
-fn scalar_size_expands_to_cube() {
-    let g = lower_src(r#"scene { box "b" (size=2) }"#);
-    let (min, max) = mesh_aabb(&g, "b");
-    assert!((min - Vec3::splat(-1.0)).abs().max_element() < 1e-5);
-    assert!((max - Vec3::splat(1.0)).abs().max_element() < 1e-5);
-}
-
-#[test]
-fn whd_shortcuts_populate_size() {
-    let g = lower_src(r#"scene { box "b" (w=2, h=4, d=6) }"#);
-    let (min, max) = mesh_aabb(&g, "b");
-    assert!((max.x - min.x - 2.0).abs() < 1e-5);
-    assert!((max.y - min.y - 4.0).abs() < 1e-5);
-    assert!((max.z - min.z - 6.0).abs() < 1e-5);
-}
-
-#[test]
-fn whd_overrides_individual_size_components() {
-    let g = lower_src(r#"scene { box "b" (size=[1, 1, 1], h=3) }"#);
-    let (min, max) = mesh_aabb(&g, "b");
-    assert!((max.x - min.x - 1.0).abs() < 1e-5);
-    assert!((max.y - min.y - 3.0).abs() < 1e-5);
-    assert!((max.z - min.z - 1.0).abs() < 1e-5);
-}
-
-#[test]
-fn xyz_shortcuts_set_translation() {
-    let g = lower_src(r#"scene { box "b" (y=1.5, size=1) }"#);
-    let t = find_mesh_node(&g, "b").transform.translation;
-    assert!((t - Vec3::new(0.0, 1.5, 0.0)).abs().max_element() < 1e-5);
-}
-
-#[test]
-fn rxyz_shortcuts_set_rotation() {
-    let g = lower_src(r#"scene { box "b" (ry=90, size=1) }"#);
-    let q = find_mesh_node(&g, "b").transform.rotation;
-    // 90° around Y rotates +X to -Z.
-    let v = q * Vec3::X;
-    assert!((v - Vec3::new(0.0, 0.0, -1.0)).abs().max_element() < 1e-4,
-        "got {v:?}");
-}
-
-#[test]
-fn anchor_bottom_places_mesh_above_origin() {
-    let g = lower_src(r#"scene { box "b" (size=2, anchor=bottom) }"#);
-    let (min, max) = mesh_aabb(&g, "b");
-    assert!(min.y.abs() < 1e-5, "expected bottom on y=0, got {min:?}");
-    assert!((max.y - 2.0).abs() < 1e-5);
-}
-
-#[test]
-fn anchor_corner_combines_axes() {
-    let g = lower_src(r#"scene { box "b" (size=2, anchor=bottom_left_front) }"#);
-    let (min, _) = mesh_aabb(&g, "b");
-    // All three mins should sit at 0.
-    assert!(min.x.abs() < 1e-5 && min.y.abs() < 1e-5 && min.z.abs() < 1e-5,
-        "expected all-mins at 0, got {min:?}");
-}
-
-#[test]
-fn anchor_shifts_default_connectors() {
-    // Anchor=bottom puts the box's bottom face on y=0; the `bottom`
-    // default connector must follow — otherwise attach math breaks.
-    let g = lower_src(r#"scene { box "b" (size=2, anchor=bottom) }"#);
-    let n = find_mesh_node(&g, "b");
-    let bottom = n.connectors.iter().find(|c| c.name == "bottom")
-        .expect("missing bottom connector");
-    assert!(bottom.pos.y.abs() < 1e-5,
-        "bottom connector should be at y=0, got {:?}", bottom.pos);
-    let top = n.connectors.iter().find(|c| c.name == "top")
-        .expect("missing top connector");
-    assert!((top.pos.y - 2.0).abs() < 1e-5,
-        "top connector should be at y=2, got {:?}", top.pos);
-}
-
-#[test]
-fn slab_defaults_to_bottom_anchor() {
-    let g = lower_src(r#"scene { slab "floor" (size=[2, 0.2, 2]) }"#);
-    let (min, max) = mesh_aabb(&g, "floor");
-    assert!(min.y.abs() < 1e-5, "slab should sit on y=0");
-    assert!((max.y - 0.2).abs() < 1e-5);
-}
-
-#[test]
-fn panel_defaults_to_back_anchor() {
-    let g = lower_src(r#"scene { panel "p" (size=[2, 2, 0.1]) }"#);
-    let (min, max) = mesh_aabb(&g, "p");
-    // Back face is the +Z face. Anchor=back means the +Z face lands at z=0.
-    assert!(max.z.abs() < 1e-5, "panel back face should be at z=0, got max.z={}", max.z);
-    assert!((min.z + 0.1).abs() < 1e-5);
-}
-
-#[test]
-fn from_to_derives_size_and_pos() {
-    let g = lower_src(r#"scene { box "b" (from=[-1, 0, -1], to=[1, 2, 1]) }"#);
-    let t = find_mesh_node(&g, "b").transform.translation;
-    assert!((t - Vec3::new(0.0, 1.0, 0.0)).abs().max_element() < 1e-5);
-    let (min, max) = mesh_aabb(&g, "b");
-    assert!((max - min - Vec3::new(2.0, 2.0, 2.0)).abs().max_element() < 1e-5);
-}
-
-#[test]
-fn stack_y_packs_children_bottom_up() {
-    let g = lower_src(
-        r#"
-        scene {
-          stack "tower" (axis=y) {
-            box "a" (size=[1, 1, 1])
-            box "b" (size=[1, 2, 1])
-            box "c" (size=[1, 0.5, 1])
-          }
-        }
-        "#,
-    );
-    let ay = find_mesh_node(&g, "a").transform.translation.y;
-    let by = find_mesh_node(&g, "b").transform.translation.y;
-    let cy = find_mesh_node(&g, "c").transform.translation.y;
-    // Each box's *center* sits at cumulative_base + half_height.
-    // a: 0 + 0.5 = 0.5; b: 1 + 1 = 2.0; c: 3 + 0.25 = 3.25.
-    assert!((ay - 0.5).abs() < 1e-4, "got a.y={ay}");
-    assert!((by - 2.0).abs() < 1e-4, "got b.y={by}");
-    assert!((cy - 3.25).abs() < 1e-4, "got c.y={cy}");
-}
-
-#[test]
-fn stack_gap_inserts_space_between_children() {
-    let g = lower_src(
-        r#"
-        scene {
-          stack "s" (axis=y, gap=0.5) {
-            box "a" (size=[1, 1, 1])
-            box "b" (size=[1, 1, 1])
-          }
-        }
-        "#,
-    );
-    let ay = find_mesh_node(&g, "a").transform.translation.y;
-    let by = find_mesh_node(&g, "b").transform.translation.y;
-    // a center at 0.5; gap of 0.5 → b center at 1 + 0.5 + 0.5 = 2.0.
-    assert!((by - ay - 1.5).abs() < 1e-4, "gap not applied: a={ay} b={by}");
-}
-
-#[test]
-fn grid_replicates_children() {
-    let g = lower_src(
-        r#"
-        scene {
-          grid "tiles" (count=[3, 1, 2], step=[1, 0, 1]) {
-            box "t" (size=[0.9, 0.1, 0.9])
-          }
-        }
-        "#,
-    );
-    // Expect 3*1*2 = 6 instance wrappers, each with a nested box.
-    let t_count = g.nodes.iter().filter(|n| n.name == "t").count();
-    assert_eq!(t_count, 6, "grid should produce 6 tiles, got {t_count}");
-}
-
-#[test]
-fn grid_attach_applies_uniformly_per_instance() {
-    // Regression: attach inside a grid body must not leak to the global
-    // resolve_attaches pass — otherwise the first instance gets attach
-    // applied twice (once globally, once per-instance) and ends up at
-    // 2× the offset of the others.
-    let g = lower_src(
-        r#"
-        scene {
-          grid "row" (count=[4, 1, 1], step=[0.5, 0, 0]) {
-            sphere "body" (radius=0.1)
-            cylinder "cap" (radius=0.05, height=0.02)
-            attach (parent="body", child="cap", socket="top", plug="bottom")
-          }
-        }
-        "#,
-    );
-    let cap_ys: Vec<f32> = g
-        .nodes
-        .iter()
-        .filter(|n| n.name == "cap")
-        .map(|n| n.transform.translation.y)
-        .collect();
-    assert_eq!(cap_ys.len(), 4, "expected 4 cap instances, got {}", cap_ys.len());
-    let first = cap_ys[0];
-    for (i, y) in cap_ys.iter().enumerate() {
-        assert!(
-            (y - first).abs() < 1e-5,
-            "cap[{i}] y={y} differs from cap[0] y={first} — attach applied unevenly across grid instances"
-        );
-    }
-}
-
-#[test]
-fn relative_placement_above_snaps_flush() {
-    let g = lower_src(
-        r#"
-        scene {
-          group "world" {
-            box "base" (size=[2, 1, 2])
-            box "hat"  (size=[1, 1, 1], above="base")
-          }
-        }
-        "#,
-    );
-    // base center y=0, top y=0.5. hat bottom flush → center at y=1.0.
-    let hat_y = find_mesh_node(&g, "hat").transform.translation.y;
-    assert!((hat_y - 1.0).abs() < 1e-4, "hat should be at y=1.0, got {hat_y}");
-}
-
-#[test]
-fn relative_placement_honors_gap() {
-    let g = lower_src(
-        r#"
-        scene {
-          group "world" {
-            box "base" (size=[2, 1, 2])
-            box "hat"  (size=[1, 1, 1], above="base", gap=0.25)
-          }
-        }
-        "#,
-    );
-    let hat_y = find_mesh_node(&g, "hat").transform.translation.y;
-    assert!((hat_y - 1.25).abs() < 1e-4, "hat should be at y=1.25, got {hat_y}");
-}
-
-#[test]
-fn explicit_pos_axis_wins_over_relative_placement() {
-    // Explicit `pos` along the placement axis must survive — without this,
-    // `behind` silently overwrites `pos.z`, which gave the imported
-    // bed-with-headboard a different headboard position than the
-    // standalone bed (top-level siblings skipped relative placement
-    // entirely, so `pos` happened to win there).
-    let g = lower_src(
-        r#"
-        scene {
-          group "world" {
-            box "base" (size=[2, 1, 2])
-            box "back" (size=[2, 1, 0.1], behind="base", pos=[0, 0, 0.75])
-          }
-        }
-        "#,
-    );
-    let z = find_mesh_node(&g, "back").transform.translation.z;
-    assert!((z - 0.75).abs() < 1e-4, "explicit pos.z should win, got {z}");
-}
-
-#[test]
-fn relative_placement_still_fires_when_pos_axis_is_zero() {
-    // Pos on a perpendicular axis must not block the snap.
-    let g = lower_src(
-        r#"
-        scene {
-          group "world" {
-            box "base" (size=[2, 1, 2])
-            box "hat"  (size=[1, 1, 1], above="base", pos=[0.25, 0, 0])
-          }
-        }
-        "#,
-    );
-    let t = find_mesh_node(&g, "hat").transform.translation;
-    // Snap on Y still fires: hat at y=1.0 (base top + hat half-height).
-    assert!((t.y - 1.0).abs() < 1e-4, "snap on Y should still fire, got {t:?}");
-    // Pos.x preserved.
-    assert!((t.x - 0.25).abs() < 1e-4, "pos.x should be preserved, got {t:?}");
-}
-
-#[test]
-fn lod_scale_halves_default_segment_count() {
-    let baseline = lower_src(r#"scene { sphere "s" (radius=0.5) }"#);
-    let scaled = lower_src(r#"lod_scale (value=0.5) scene { sphere "s" (radius=0.5) }"#);
-    let base_verts = find_mesh_node(&baseline, "s").mesh.as_ref().unwrap().positions.len();
-    let scaled_verts = find_mesh_node(&scaled, "s").mesh.as_ref().unwrap().positions.len();
-    assert!(
-        scaled_verts < base_verts,
-        "lod_scale=0.5 should reduce sphere vert count (base={base_verts}, scaled={scaled_verts})"
-    );
-}
-
-#[test]
-fn lod_scale_doubles_default_segment_count() {
-    let baseline = lower_src(r#"scene { cylinder "c" (radius=0.5, height=1) }"#);
-    let scaled = lower_src(
-        r#"lod_scale (value=2) scene { cylinder "c" (radius=0.5, height=1) }"#,
-    );
-    let base_verts = find_mesh_node(&baseline, "c").mesh.as_ref().unwrap().positions.len();
-    let scaled_verts = find_mesh_node(&scaled, "c").mesh.as_ref().unwrap().positions.len();
-    assert!(
-        scaled_verts > base_verts,
-        "lod_scale=2 should increase cylinder vert count (base={base_verts}, scaled={scaled_verts})"
-    );
-}
-
-#[test]
-fn lod_scale_also_scales_explicit_segments() {
-    // Explicit per-primitive segment counts ride the global multiplier
-    // too — otherwise dense surface primitives (heightfield, curved_plane,
-    // surf-style waves) become LOD-inert. The clamp keeps the count above
-    // each primitive's minimum so circles still close.
-    let baseline = lower_src(r#"scene { sphere "s" (radius=0.5, rings=16, segments=24) }"#);
-    let scaled = lower_src(
-        r#"lod_scale (value=0.25) scene { sphere "s" (radius=0.5, rings=16, segments=24) }"#,
-    );
-    let base_verts = find_mesh_node(&baseline, "s").mesh.as_ref().unwrap().positions.len();
-    let scaled_verts = find_mesh_node(&scaled, "s").mesh.as_ref().unwrap().positions.len();
-    assert!(
-        scaled_verts < base_verts,
-        "lod_scale=0.25 should reduce explicit rings=16/segments=24 (base={base_verts}, scaled={scaled_verts})"
-    );
-}
-
-#[test]
-fn lod_scale_scales_explicit_heightfield_grid() {
-    // Regression for examples/nature/heightfield_terrain.mog and wave_water.mog —
-    // both pin segments_u/segments_v and would otherwise ignore lod_scale.
-    let baseline = lower_src(
-        r#"scene { heightfield "h" (size=[6, 6], segments_u=64, segments_v=64) }"#,
-    );
-    let halved = lower_src(
-        r#"lod_scale (value=0.5) scene { heightfield "h" (size=[6, 6], segments_u=64, segments_v=64) }"#,
-    );
-    let doubled = lower_src(
-        r#"lod_scale (value=2) scene { heightfield "h" (size=[6, 6], segments_u=64, segments_v=64) }"#,
-    );
-    let verts = |g: &SceneGraph| find_mesh_node(g, "h").mesh.as_ref().unwrap().positions.len();
-    assert!(verts(&halved) < verts(&baseline));
-    assert!(verts(&doubled) > verts(&baseline));
-}
-
-#[test]
-fn lod_scale_clamps_explicit_segments_to_min() {
-    // A vanishing lod_scale must not silently destroy a primitive — every
-    // segment count is floored at its per-primitive minimum (cylinder=3).
-    let g = lower_src(
-        r#"lod_scale (value=0.01) scene { cylinder "c" (radius=0.5, height=1, segments=24) }"#,
-    );
-    let mesh = find_mesh_node(&g, "c").mesh.as_ref().unwrap();
-    // Cylinder side ring + caps; with segments=3 it's still a closed solid.
-    assert!(
-        !mesh.positions.is_empty() && !mesh.indices.is_empty(),
-        "cylinder under lod_scale=0.01 must still produce geometry"
-    );
-}
-
-#[test]
-fn lod_scale_steps_icosphere_subdivisions() {
-    // Icosphere triangle count is 20 * 4^subdivisions. Default subdivisions=2
-    // → 320 tris. lod_scale=2 → subdivisions=3 → 1280 tris. lod_scale=0.5 →
-    // subdivisions=1 → 80 tris.
-    let base = lower_src(r#"scene { icosphere "i" (radius=0.5) }"#);
-    let up = lower_src(r#"lod_scale (value=2) scene { icosphere "i" (radius=0.5) }"#);
-    let down = lower_src(r#"lod_scale (value=0.5) scene { icosphere "i" (radius=0.5) }"#);
-    let tris = |g: &SceneGraph| find_mesh_node(g, "i").mesh.as_ref().unwrap().indices.len() / 3;
-    assert_eq!(tris(&base), 320);
-    assert_eq!(tris(&up), 1280);
-    assert_eq!(tris(&down), 80);
-}
-
-#[test]
-fn lod_scale_default_keeps_existing_vertex_counts() {
-    // No `lod_scale` directive should leave every default mesh untouched —
-    // a regression here would silently change every existing .mog's output.
-    let g = lower_src(r#"scene { sphere "s" (radius=0.5) cylinder "c" (radius=0.5, height=1) }"#);
-    let s_verts = find_mesh_node(&g, "s").mesh.as_ref().unwrap().positions.len();
-    let c_verts = find_mesh_node(&g, "c").mesh.as_ref().unwrap().positions.len();
-    // Sphere default rings=16, segments=24 → 17 * 25 = 425 verts (one extra
-    // ring + one extra segment for the seam); cylinder default segments=24
-    // → 2 * (24 + 1) side verts + 2 * (24 + 1) cap-fan verts + 2 cap centres
-    // = 102 verts. These exact counts depend on the mesh builder — the
-    // test asserts them so a future LOD-scale change doesn't drift defaults.
-    assert_eq!(s_verts, 425);
-    assert_eq!(c_verts, 102);
-}
-
-#[test]
-fn per_node_lod_doubles_segment_count_on_marked_subtree() {
-    // `lod=2.0` on a single primitive doubles its default segment count
-    // (matches the behaviour of `lod_scale (value=2)` but scoped to that
-    // subtree only — see lod.rs::LodMultiplierGuard).
-    let baseline = lower_src(r#"scene { cylinder "c" (radius=0.5, height=1) }"#);
-    let scaled = lower_src(r#"scene { cylinder "c" (radius=0.5, height=1, lod=2) }"#);
-    let base_verts = find_mesh_node(&baseline, "c").mesh.as_ref().unwrap().positions.len();
-    let scaled_verts = find_mesh_node(&scaled, "c").mesh.as_ref().unwrap().positions.len();
-    assert!(
-        scaled_verts > base_verts,
-        "per-node lod=2 should increase cylinder vert count (base={base_verts}, scaled={scaled_verts})"
-    );
-}
-
-#[test]
-fn per_node_lod_does_not_leak_to_siblings() {
-    // The multiplier guard is RAII-scoped to the marked subtree. A `lod=2`
-    // group must not boost a sibling that lives outside it.
-    let g = lower_src(
-        r#"scene {
-            group "hi" (lod=2) { sphere "s" (radius=0.5) }
-            sphere "lo" (radius=0.5)
-        }"#,
-    );
-    let hi = find_mesh_node(&g, "s").mesh.as_ref().unwrap().positions.len();
-    let lo = find_mesh_node(&g, "lo").mesh.as_ref().unwrap().positions.len();
-    let baseline = lower_src(r#"scene { sphere "b" (radius=0.5) }"#);
-    let base = find_mesh_node(&baseline, "b").mesh.as_ref().unwrap().positions.len();
-    assert!(hi > base, "lod=2 group should boost child sphere (hi={hi}, base={base})");
-    assert_eq!(lo, base, "sibling outside the lod=2 group must use baseline detail");
-}
-
-#[test]
-fn per_node_lod_compounds_with_global_lod_scale() {
-    // `lod=2.0` on top of `lod_scale (value=0.5)` cancels out — effective
-    // multiplier is 1.0, so the marked subtree returns to the default
-    // vertex count even though the file's global setting is 0.5.
-    let baseline = lower_src(r#"scene { sphere "s" (radius=0.5) }"#);
-    let compound = lower_src(
-        r#"lod_scale (value=0.5) scene { sphere "s" (radius=0.5, lod=2) }"#,
-    );
-    let base_verts = find_mesh_node(&baseline, "s").mesh.as_ref().unwrap().positions.len();
-    let compound_verts = find_mesh_node(&compound, "s").mesh.as_ref().unwrap().positions.len();
-    assert_eq!(
-        base_verts, compound_verts,
-        "lod=2 should cancel a global lod_scale=0.5 (base={base_verts}, compound={compound_verts})"
-    );
-}
-
-#[test]
-fn extrude_lowers_with_default_attrs() {
-    let g = lower_src(r#"scene { extrude "block" () }"#);
-    let m = find_mesh_node(&g, "block").mesh.as_ref().unwrap();
-    assert!(!m.positions.is_empty(), "extrude default mesh missing positions");
-    assert!(!m.indices.is_empty(), "extrude default mesh missing indices");
-    // Default outline is a unit square 1×1 in XZ; default height 1.
-    let (min, max) = mesh_aabb(&g, "block");
-    assert!((max.y - min.y - 1.0).abs() < 1e-3, "Y span should be 1");
-}
-
-#[test]
-fn extrude_with_explicit_polygon_shape() {
-    // L-shaped polygon: not convex, exercises earcut.
-    let g = lower_src(
-        r#"scene {
-            extrude "ell" (
-                points=[[0, 0], [2, 0], [2, 1], [1, 1], [1, 2], [0, 2]],
-                height=0.5
-            )
-        }"#,
-    );
-    let (_min, max) = mesh_aabb(&g, "ell");
-    assert!((max.x - 2.0).abs() < 1e-3 && (max.z - 2.0).abs() < 1e-3,
-        "L-shape should span 2 units on X and Z");
-    // Concave triangulation must still emit non-trivial tri count — earcut
-    // on a 6-vert poly emits 4 cap tris × 2 caps + 12 side rib tris.
-    let m = find_mesh_node(&g, "ell").mesh.as_ref().unwrap();
-    assert!(m.indices.len() / 3 > 8, "concave extrude should emit > 8 tris");
-}
-
-#[test]
-fn sweep_lowers_along_path() {
-    let g = lower_src(
-        r#"scene {
-            sweep "rail" (
-                profile=[[-0.05, -0.02], [0.05, -0.02], [0.05, 0.02], [-0.05, 0.02]],
-                path=[[-1, 0, 0], [0, 0, 0], [1, 0, 0]],
-                samples=8
-            )
-        }"#,
-    );
-    let (min, max) = mesh_aabb(&g, "rail");
-    assert!(min.x < -0.9 && max.x > 0.9, "swept rail should span ±1 along path");
-    let m = find_mesh_node(&g, "rail").mesh.as_ref().unwrap();
-    assert!(!m.indices.is_empty(), "sweep produced no triangles");
-}
-
-#[test]
-fn loft_lowers_three_sections() {
-    let g = lower_src(
-        r#"scene {
-            loft "hull" (
-                points=[
-                    [-0.5, -0.2], [0.5, -0.2], [0.5, 0.2], [-0.5, 0.2],
-                    [-1.0, -0.4], [1.0, -0.4], [1.0, 0.4], [-1.0, 0.4],
-                    [-0.6, -0.1], [0.6, -0.1], [0.6, 0.1], [-0.6, 0.1]
-                ],
-                heights=[0.0, 1.0, 2.0]
-            )
-        }"#,
-    );
-    let (min, max) = mesh_aabb(&g, "hull");
-    assert!(min.y.abs() < 1e-3 && (max.y - 2.0).abs() < 1e-3,
-        "Y span should be [0, 2] (got [{}, {}])", min.y, max.y);
-    // Middle section reaches ±1 on X.
-    assert!(max.x > 0.95, "mid section X extent should reach ~1 (got {})", max.x);
-}
-
-#[test]
-fn loft_rejects_inconsistent_section_lengths() {
-    // points contains 5 vertices, heights demands 2 sections — 5 % 2 ≠ 0.
-    let src = r#"scene {
-        loft "bad" (
-            points=[[0, 0], [1, 0], [1, 1], [0, 1], [0.5, 0.5]],
-            heights=[0.0, 1.0]
-        )
-    }"#;
-    let ast = crate::parser::parse(src).unwrap();
-    assert!(lower(&ast).is_err(), "loft must reject mismatched section lengths");
-}
-
-#[test]
-fn loft_rejects_single_section() {
-    // heights with fewer than 2 entries leaves nothing to interpolate
-    // between — must error rather than emit a degenerate mesh.
-    let src = r#"scene {
-        loft "bad" (
-            points=[[0, 0], [1, 0], [1, 1], [0, 1]],
-            heights=[0.0]
-        )
-    }"#;
-    let ast = crate::parser::parse(src).unwrap();
-    assert!(lower(&ast).is_err(), "loft must reject heights.len() < 2");
-}
-
-#[test]
-fn loft_rejects_section_with_fewer_than_three_vertices() {
-    // 2 sections × 2 verts per section → per_section=2 < 3 (a loft section
-    // needs to be a closed polygon with at least 3 vertices).
-    let src = r#"scene {
-        loft "bad" (
-            points=[[0, 0], [1, 0], [0, 1], [1, 1]],
-            heights=[0.0, 1.0]
-        )
-    }"#;
-    let ast = crate::parser::parse(src).unwrap();
-    assert!(lower(&ast).is_err(), "loft must reject per-section vertex count < 3");
-}
-
-#[test]
-fn light_directional_lowers_with_color_and_intensity() {
-    let g = lower_src(
-        r#"scene { light "sun" (kind=directional, color=[1, 0.95, 0.85], intensity=3) }"#,
-    );
-    let n = find_mesh_node(&g, "sun");
-    assert!(n.mesh.is_none(), "light should not carry a mesh");
-    let l = n.light.as_ref().expect("light field set");
-    assert_eq!(l.kind, mogen_core::LightKind::Directional);
-    assert_eq!(l.color, [1.0, 0.95, 0.85]);
-    assert!((l.intensity - 3.0).abs() < 1e-6);
-    assert!(l.range.is_none());
-}
-
-#[test]
-fn light_point_carries_range() {
-    let g = lower_src(
-        r#"scene { light "lamp" (kind=point, pos=[0, 2, 0], intensity=10, range=8) }"#,
-    );
-    let n = find_mesh_node(&g, "lamp");
-    let l = n.light.as_ref().unwrap();
-    assert_eq!(l.kind, mogen_core::LightKind::Point);
-    assert_eq!(l.range, Some(8.0));
-    assert!((n.transform.translation - Vec3::new(0.0, 2.0, 0.0)).abs().max_element() < 1e-5);
-}
-
-#[test]
-fn light_spot_converts_cone_degrees_to_radians() {
-    let g = lower_src(
-        r#"scene { light "spot" (kind=spot, intensity=20, range=10, inner_cone=20, outer_cone=35) }"#,
-    );
-    let l = find_mesh_node(&g, "spot").light.as_ref().unwrap();
-    assert_eq!(l.kind, mogen_core::LightKind::Spot);
-    assert!((l.inner_cone_rad - 20f32.to_radians()).abs() < 1e-5);
-    assert!((l.outer_cone_rad - 35f32.to_radians()).abs() < 1e-5);
-}
-
-#[test]
-fn light_dir_synthesizes_rotation_from_neg_z() {
-    // dir=[0,-1,0] should rotate the node so its local -Z points down.
-    let g = lower_src(
-        r#"scene { light "sun" (kind=directional, dir=[0, -1, 0]) }"#,
-    );
-    let n = find_mesh_node(&g, "sun");
-    let forward = n.transform.rotation * Vec3::NEG_Z;
-    assert!(
-        (forward - Vec3::new(0.0, -1.0, 0.0)).length() < 1e-5,
-        "expected -Z to map to (0,-1,0), got {forward:?}"
-    );
-}
-
-#[test]
-fn wall_cuts_holes_via_csg() {
-    // A wall with one hole should have substantially fewer closed-face
-    // verts on its interior than a plain box would, and the hole should
-    // leave a gap in the y-z cross section at x=0.
-    let g = lower_src(
-        r#"scene { wall "w" (size=[4, 3, 0.1], holes=[[0, 0, 1, 2]]) }"#,
-    );
-    let mesh = find_mesh_node(&g, "w").mesh.as_ref().unwrap();
-    // No vertex should lie strictly inside the hole rectangle on the
-    // front/back face (x in [-0.5, 0.5], y in [-1, 1], z ~ 0).
-    // Instead, the hole boundary sits at x=±0.5 / y=±1, which is fine.
-    let strict_inside = mesh.positions.iter().any(|p| {
-        p[0].abs() < 0.45 && p[1].abs() < 0.95 && (p[2].abs() - 0.05).abs() < 0.01
-    });
-    assert!(!strict_inside, "wall hole interior should be empty");
-}
-
-#[test]
-fn deform_noise_changes_mesh_positions() {
-    // Same primitive built once plain and once with noise+seed should differ.
-    let plain = lower_src(r#"scene { sphere "s" (radius=0.5) }"#);
-    let noisy = lower_src(r#"scene { sphere "s" (radius=0.5, noise=0.4, seed=7) }"#);
-    let p = find_mesh_node(&plain, "s").mesh.as_ref().unwrap();
-    let n = find_mesh_node(&noisy, "s").mesh.as_ref().unwrap();
-    // Auto-bumped tessellation: noisy mesh should have more verts than plain.
-    assert!(
-        n.positions.len() > p.positions.len(),
-        "expected noise to bump tessellation: plain={}, noisy={}",
-        p.positions.len(),
-        n.positions.len()
-    );
-    // Per-vertex positions necessarily differ from the plain unit sphere.
-    let plain_radii: f32 = p.positions.iter()
-        .map(|q| (q[0].powi(2) + q[1].powi(2) + q[2].powi(2)).sqrt())
-        .sum::<f32>() / p.positions.len() as f32;
-    let noisy_radii: f32 = n.positions.iter()
-        .map(|q| (q[0].powi(2) + q[1].powi(2) + q[2].powi(2)).sqrt())
-        .sum::<f32>() / n.positions.len() as f32;
-    // Plain sphere radii average ~0.5; noise perturbs them but the mean stays
-    // near 0.5 (zero-mean random shift). What we want is that the per-vertex
-    // STD is non-trivial.
-    let noisy_std: f32 = (n.positions.iter()
-        .map(|q| {
-            let r = (q[0].powi(2) + q[1].powi(2) + q[2].powi(2)).sqrt();
-            (r - noisy_radii).powi(2)
-        })
-        .sum::<f32>() / n.positions.len() as f32).sqrt();
-    let plain_std: f32 = (p.positions.iter()
-        .map(|q| {
-            let r = (q[0].powi(2) + q[1].powi(2) + q[2].powi(2)).sqrt();
-            (r - plain_radii).powi(2)
-        })
-        .sum::<f32>() / p.positions.len() as f32).sqrt();
-    assert!(
-        noisy_std > plain_std + 0.001,
-        "expected noise to widen radius distribution: plain_std={plain_std}, noisy_std={noisy_std}"
-    );
-}
-
-#[test]
-fn deform_seed_determinism() {
-    // Same source compiled twice should produce byte-identical positions.
-    let a = lower_src(r#"scene { box "b" (size=[1,1,1], noise=0.3, seed=42) }"#);
-    let b = lower_src(r#"scene { box "b" (size=[1,1,1], noise=0.3, seed=42) }"#);
-    let pa = find_mesh_node(&a, "b").mesh.as_ref().unwrap();
-    let pb = find_mesh_node(&b, "b").mesh.as_ref().unwrap();
-    assert_eq!(pa.positions, pb.positions);
-}
-
-#[test]
-fn deform_taper_shrinks_top_of_cylinder() {
-    let g = lower_src(
-        r#"scene { cylinder "c" (radius=0.5, height=1.0, taper=0.5) }"#,
-    );
-    let mesh = find_mesh_node(&g, "c").mesh.as_ref().unwrap();
-    let mut top_max = 0.0_f32;
-    let mut bot_max = 0.0_f32;
-    for p in &mesh.positions {
-        let r = (p[0] * p[0] + p[2] * p[2]).sqrt();
-        if p[1] > 0.4 { top_max = top_max.max(r); }
-        if p[1] < -0.4 { bot_max = bot_max.max(r); }
-    }
-    assert!((top_max - 0.25).abs() < 1e-3, "top radius should be ~0.25, got {top_max}");
-    assert!((bot_max - 0.5).abs() < 1e-3, "bottom radius should be ~0.5, got {bot_max}");
-}
-
-#[test]
-fn mirror_bakes_reflection_into_subtree_so_chain_stays_positive_det() {
-    // Regression: a `mirror (axis=x)` used to leave its second instance with
-    // a `scale=(-1,1,1)` on the wrapper, which gave it a negative-determinant
-    // world transform. Renderers that don't reverse front-face winding for
-    // negative-det chains (and most glTF importers in practice) drew the
-    // mirrored copy backface-culled — the `gaming_chair.mog` armpad bug.
-    let g = lower_src(
-        r#"
-        scene {
-          mirror "pair" (axis=x) {
-            box "leaf" (pos=[0.5, 0.0, 0.0], size=[0.2, 0.2, 0.2])
-          }
-        }
-        "#,
-    );
-
-    // Every node's local scale must be positive after the bake.
-    for n in &g.nodes {
-        let s = n.transform.scale;
-        assert!(
-            s.x * s.y * s.z > 0.0,
-            "node `{}` has non-positive-det scale {:?} after mirror bake",
-            n.name,
-            s
-        );
-    }
-
-    // Find the original (`pair_0`) and mirrored (`pair_1`) leaves and confirm
-    // the second one has its translation flipped on x and its mesh winding
-    // reversed relative to the first.
-    let pair_0_leaf = g
-        .nodes
-        .iter()
-        .find(|n| n.name == "leaf"
-            && n.parent.is_some()
-            && g.nodes[n.parent.unwrap().0 as usize].name == "pair_0")
-        .expect("pair_0/leaf");
-    let pair_1_leaf = g
-        .nodes
-        .iter()
-        .find(|n| n.name == "leaf"
-            && n.parent.is_some()
-            && g.nodes[n.parent.unwrap().0 as usize].name == "pair_1")
-        .expect("pair_1/leaf");
-
-    assert!((pair_0_leaf.transform.translation.x - 0.5).abs() < 1e-5);
-    assert!((pair_1_leaf.transform.translation.x + 0.5).abs() < 1e-5);
-
-    let m0 = pair_0_leaf.mesh.as_ref().unwrap();
-    let m1 = pair_1_leaf.mesh.as_ref().unwrap();
-    assert_eq!(m0.indices.len(), m1.indices.len());
-    // Per-triangle winding flipped: m1 swaps indices 1 and 2 of every tri
-    // (and x-flips positions/normals).
-    for (a, b) in m0.indices.chunks_exact(3).zip(m1.indices.chunks_exact(3)) {
-        assert_eq!(a[0], b[0]);
-        assert_eq!(a[1], b[2]);
-        assert_eq!(a[2], b[1]);
-    }
-    for (p0, p1) in m0.positions.iter().zip(m1.positions.iter()) {
-        assert!((p0[0] + p1[0]).abs() < 1e-5, "x should be negated");
-        assert!((p0[1] - p1[1]).abs() < 1e-5);
-        assert!((p0[2] - p1[2]).abs() < 1e-5);
-    }
-    for (n0, n1) in m0.normals.iter().zip(m1.normals.iter()) {
-        assert!((n0[0] + n1[0]).abs() < 1e-5, "normal x should be negated");
-    }
-}
-
-#[test]
-fn mirror_flip_bind_swaps_lr_suffix_for_mirrored_copy() {
-    // `mirror (axis=x, flip_bind=1)` over a mesh bound to `shoulder_l`
-    // should produce two scene nodes, the original still bound to
-    // `shoulder_l` and the mirrored copy rebound to `shoulder_r`. The flag
-    // exists so symmetric humanoid accessories (sleeves, cuffs, shoes) can
-    // be authored once and follow both bones — string-typed module params
-    // aren't supported, so without `flip_bind` you'd hand-author each side.
-    let g = lower_src(
-        r#"
-        scene {
-          skeleton "rig" {
-            bone "shoulder_l" (pos=[ 0.2, 1.4, 0])
-            bone "shoulder_r" (pos=[-0.2, 1.4, 0])
-          }
-          mirror "arms" (axis=x, flip_bind=1) {
-            box "sleeve" (pos=[0.2, 1.4, 0], size=[0.1, 0.2, 0.1],
-                          skin="rig", bind="shoulder_l")
-          }
-        }
-        "#,
-    );
-
-    let l_idx = g.find_node("shoulder_l").unwrap().0 as u16;
-    let r_idx = g.find_node("shoulder_r").unwrap().0 as u16;
-    let skin = g.find_skin("rig").unwrap();
-    let l_joint = g.skins[skin.0 as usize]
-        .joints
-        .iter()
-        .position(|j| j.0 as u16 == l_idx)
-        .unwrap() as u16;
-    let r_joint = g.skins[skin.0 as usize]
-        .joints
-        .iter()
-        .position(|j| j.0 as u16 == r_idx)
-        .unwrap() as u16;
-
-    let sleeves: Vec<&mogen_core::SceneNode> = g
-        .nodes
-        .iter()
-        .filter(|n| n.name == "sleeve")
-        .collect();
-    assert_eq!(sleeves.len(), 2, "mirror should produce two `sleeve` copies");
-
-    // Both copies are skinned. The unmirrored copy binds to shoulder_l;
-    // the mirrored copy binds to shoulder_r via flip_bind.
-    let mut bound_to_l = 0;
-    let mut bound_to_r = 0;
-    for s in &sleeves {
-        let m = s.mesh.as_ref().expect("sleeve mesh");
-        assert_eq!(s.skin, Some(skin), "sleeve should be skinned");
-        for j in &m.joints {
-            assert_eq!(j[1..], [0, 0, 0], "rigid bind expected: only slot 0 used");
-            if j[0] == l_joint {
-                bound_to_l += 1;
-            } else if j[0] == r_joint {
-                bound_to_r += 1;
-            } else {
-                panic!("unexpected joint index {} in sleeve", j[0]);
-            }
-        }
-    }
-    assert!(bound_to_l > 0, "unmirrored copy must still bind to shoulder_l");
-    assert!(bound_to_r > 0, "mirrored copy must rebind to shoulder_r");
-}
-
-#[test]
-fn bend_z_range_only_bends_tip() {
-    // Cylinder from y=-0.5 to y=0.5; bend_z bends along Y. With range
-    // [0.75, 1.0] the lower 75 % of the column stays at x≈0; only the upper
-    // 25 % gets perturbed. Compare against the unranged form so we can see
-    // the lower ring stayed put while the upper ring moved.
-    let baseline = lower_src(
-        r#"scene { cylinder "c" (radius=0.1, height=1.0, segments=8) }"#,
-    );
-    let ranged = lower_src(
-        r#"scene { cylinder "c" (radius=0.1, height=1.0, segments=8, bend_z=60, bend_z_range=[0.75, 1.0]) }"#,
-    );
-    let base_mesh = find_mesh_node(&baseline, "c").mesh.as_ref().unwrap();
-    let bent_mesh = find_mesh_node(&ranged, "c").mesh.as_ref().unwrap();
-    assert_eq!(base_mesh.positions.len(), bent_mesh.positions.len());
-    let mut max_base_shift = 0.0_f32;
-    let mut max_tip_shift = 0.0_f32;
-    for (b, p) in base_mesh.positions.iter().zip(bent_mesh.positions.iter()) {
-        let dx = p[0] - b[0];
-        let dy = p[1] - b[1];
-        let dz = p[2] - b[2];
-        let shift = (dx * dx + dy * dy + dz * dz).sqrt();
-        // Use the unbent y to bucket: vertices originally below the column
-        // midpoint are "base"; above are "tip". (After bending the tip
-        // slides downward, so post-bend `p[1]` would mis-bucket.)
-        if b[1] < 0.0 {
-            max_base_shift = max_base_shift.max(shift);
-        } else {
-            max_tip_shift = max_tip_shift.max(shift);
-        }
-    }
-    assert!(
-        max_base_shift < 1e-3,
-        "lower half should stay put with smoothstep ramp at 0.75 (got {max_base_shift})"
-    );
-    assert!(
-        max_tip_shift > 0.3,
-        "upper half should bend appreciably (got {max_tip_shift})"
-    );
-}
-
-#[test]
-fn taper_range_leaves_lower_half_unscaled() {
-    // Sphere has dense Y rings, so we can probe both the unranged taper=0.5
-    // result and the ranged form at the same y. With taper_range=[0.5, 1.0]
-    // the lower hemisphere stays pristine (weight=0) and the upper
-    // hemisphere ramps in via smoothstep.
-    let plain = lower_src(r#"scene { sphere "s" (radius=0.5) }"#);
-    let ranged = lower_src(
-        r#"scene { sphere "s" (radius=0.5, taper=0.5, taper_range=[0.5, 1.0]) }"#,
-    );
-    let p = find_mesh_node(&plain, "s").mesh.as_ref().unwrap();
-    let r = find_mesh_node(&ranged, "s").mesh.as_ref().unwrap();
-    assert_eq!(p.positions.len(), r.positions.len());
-
-    let mut max_lower_xz_diff = 0.0_f32;
-    let mut max_upper_xz_diff = 0.0_f32;
-    for (pl, ra) in p.positions.iter().zip(r.positions.iter()) {
-        let xz_diff = ((ra[0] - pl[0]).powi(2) + (ra[2] - pl[2]).powi(2)).sqrt();
-        if pl[1] < -0.05 {
-            max_lower_xz_diff = max_lower_xz_diff.max(xz_diff);
-        } else if pl[1] > 0.3 {
-            max_upper_xz_diff = max_upper_xz_diff.max(xz_diff);
-        }
-    }
-    assert!(
-        max_lower_xz_diff < 1e-4,
-        "lower hemisphere should match the un-tapered sphere (got {max_lower_xz_diff})"
-    );
-    assert!(
-        max_upper_xz_diff > 0.05,
-        "upper hemisphere should be tapered noticeably (got {max_upper_xz_diff})"
-    );
-}
-
-#[test]
-fn noise_range_only_roughens_top() {
-    // Tall box; with noise_range=[0.7, 1.0] only the top 30 % gets bumpy.
-    // The bottom face vertices lie at y=-0.5; their (x, z) should match the
-    // pristine box — within float noise of zero.
-    let g = lower_src(
-        r#"scene { box "b" (size=[1, 1, 1], noise=0.5, seed=11, noise_range=[0.7, 1.0]) }"#,
-    );
-    let mesh = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    let mut max_base_dx = 0.0_f32;
-    let mut max_top_dx = 0.0_f32;
-    for p in &mesh.positions {
-        // Box default has corners at ±0.5 on every axis; subtracting the
-        // sign-aware nominal corner gives the displacement magnitude.
-        let dx = p[0].abs() - 0.5;
-        let dz = p[2].abs() - 0.5;
-        let max_xy = dx.abs().max(dz.abs());
-        if p[1] < -0.4 {
-            max_base_dx = max_base_dx.max(max_xy);
-        } else if p[1] > 0.4 {
-            max_top_dx = max_top_dx.max(max_xy);
-        }
-    }
-    assert!(
-        max_base_dx < 1e-3,
-        "bottom face should stay at exactly ±0.5 (got {max_base_dx})"
-    );
-    assert!(
-        max_top_dx > 0.01,
-        "top face should be perturbed by noise (got {max_top_dx})"
-    );
-}
-
-#[test]
-fn range_reversed_endpoints_are_normalised() {
-    // [1.0, 0.5] is a user typo for [0.5, 1.0]. set_range sorts the pair so
-    // the kernel sees a soft ramp instead of a hard step at 1.0 (which would
-    // give zero deformation everywhere because t < 1.0 everywhere except
-    // the topmost row).
-    let a = lower_src(
-        r#"scene { cylinder "c" (radius=0.1, height=1.0, segments=8, bend_z=60, bend_z_range=[0.5, 1.0]) }"#,
-    );
-    let b = lower_src(
-        r#"scene { cylinder "c" (radius=0.1, height=1.0, segments=8, bend_z=60, bend_z_range=[1.0, 0.5]) }"#,
-    );
-    let pa = find_mesh_node(&a, "c").mesh.as_ref().unwrap();
-    let pb = find_mesh_node(&b, "c").mesh.as_ref().unwrap();
-    assert_eq!(pa.positions.len(), pb.positions.len());
-    for (p, q) in pa.positions.iter().zip(pb.positions.iter()) {
-        for k in 0..3 {
-            assert!(
-                (p[k] - q[k]).abs() < 1e-5,
-                "swapped endpoints produced different geometry at axis {k}"
-            );
-        }
-    }
-}
-#[test]
-fn chamfered_box_lowers_with_default_attrs() {
-    let g = lower_src(r#"scene { chamfered_box "b" () }"#);
-    let m = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    // 6 face rects (12 tris) + 12 bevel quads (24 tris) + 8 corner tris (8) = 44.
-    assert_eq!(m.indices.len() / 3, 44);
-    // Default size is 1×1×1 so the AABB sits in [-0.5, 0.5] on every axis.
-    let mut mn = [f32::INFINITY; 3];
-    let mut mx = [f32::NEG_INFINITY; 3];
-    for p in &m.positions {
-        for k in 0..3 {
-            mn[k] = mn[k].min(p[k]);
-            mx[k] = mx[k].max(p[k]);
-        }
-    }
-    for k in 0..3 {
-        assert!((mx[k] - 0.5).abs() < 1e-5, "axis {k} mx {} not 0.5", mx[k]);
-        assert!((mn[k] + 0.5).abs() < 1e-5, "axis {k} mn {} not -0.5", mn[k]);
-    }
-}
-
-#[test]
-fn chamfered_box_radius_zero_matches_plain_box() {
-    let chamfered = lower_src(r#"scene { chamfered_box "b" (radius=0) }"#);
-    let plain = lower_src(r#"scene { box "b" () }"#);
-    let cm = find_mesh_node(&chamfered, "b").mesh.as_ref().unwrap();
-    let pm = find_mesh_node(&plain, "b").mesh.as_ref().unwrap();
-    assert_eq!(cm.positions.len(), pm.positions.len());
-    assert_eq!(cm.indices.len(), pm.indices.len());
-}
-
-#[test]
-fn inset_box_lowers_with_default_face() {
-    let g = lower_src(r#"scene { inset_box "b" (size=[1, 1, 1]) }"#);
-    let m = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    // Default face is +y; the sunken floor should sit at y = 0.5 - 0.05 = 0.45.
-    let mut floor_y_seen = false;
-    for p in &m.positions {
-        if (p[1] - 0.45).abs() < 1e-5 {
-            floor_y_seen = true;
-            break;
-        }
-    }
-    assert!(floor_y_seen, "expected a vertex at the default sunken Y=0.45");
-}
-
-#[test]
-fn inset_box_face_aliases_resolve() {
-    // "top" must resolve to +y; the resulting mesh should be identical
-    // to the canonical "+y" form vertex-for-vertex.
-    let a = lower_src(r#"scene { inset_box "b" (face="+y", amount=0.2, depth=0.1) }"#);
-    let b = lower_src(r#"scene { inset_box "b" (face="top", amount=0.2, depth=0.1) }"#);
-    let pa = find_mesh_node(&a, "b").mesh.as_ref().unwrap();
-    let pb = find_mesh_node(&b, "b").mesh.as_ref().unwrap();
-    assert_eq!(pa.positions, pb.positions);
-    assert_eq!(pa.indices, pb.indices);
-}
-
-#[test]
-fn inset_box_unknown_face_errors() {
-    let ast = parse(r#"scene { inset_box "b" (face="diagonal") }"#).expect("parse");
-    let err = lower(&ast).expect_err("unknown face must error");
-    let msg = format!("{err:#}");
-    assert!(
-        msg.contains("inset_box.face") && msg.contains("diagonal"),
-        "error should name the bad face value, got: {msg}"
-    );
-}
-
-#[test]
-fn inset_box_minus_x_face_sinks_along_negative_x() {
-    // face="-x" should produce a sunken floor at x = -hx + depth = -0.5 + 0.1 = -0.4.
-    let g = lower_src(r#"scene { inset_box "b" (face="-x", amount=0.2, depth=0.1) }"#);
-    let m = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    let mut floor_x_seen = false;
-    for p in &m.positions {
-        if (p[0] + 0.4).abs() < 1e-5 {
-            floor_x_seen = true;
-            break;
-        }
-    }
-    assert!(floor_x_seen, "expected vertex at sunken X=-0.4");
-}
-// ─── Phase C3: control flow + string interpolation ──────────────────────
-
-#[test]
-fn for_loop_emits_n_copies_with_loop_var_in_pos() {
-    // Outside of a module, `for` still works because expand_modules walks
-    // the top-level child list through expand_children_into.
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=0, to=4) {
-                box "leg" (size=[0.05, 0.5, 0.05], pos=[$i * 0.3, 0, 0])
-            }
-        }"#,
-    );
-    let legs: Vec<_> = g.nodes.iter().filter(|n| n.name == "leg").collect();
-    assert_eq!(legs.len(), 4, "for(0..4) should emit 4 nodes");
-    let xs: Vec<f32> = legs.iter().map(|n| n.transform.translation.x).collect();
-    let mut want = vec![0.0, 0.3, 0.6, 0.9];
-    let mut got = xs.clone();
-    want.sort_by(|a, b| a.partial_cmp(b).unwrap());
-    got.sort_by(|a, b| a.partial_cmp(b).unwrap());
-    for (a, b) in want.iter().zip(got.iter()) {
-        assert!((a - b).abs() < 1e-5, "expected x={a}, got {b}");
-    }
-}
-
-#[test]
-fn for_loop_step_2_skips_odd_indices() {
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=0, to=6, step=2) {
-                box "even_$i" (size=[0.1, 0.1, 0.1])
-            }
-        }"#,
-    );
-    let names: Vec<_> = g.nodes.iter().filter(|n| n.name.starts_with("even_")).map(|n| n.name.clone()).collect();
-    assert_eq!(names.len(), 3, "0,2,4 = 3 iterations");
-    assert!(names.contains(&"even_0".to_string()));
-    assert!(names.contains(&"even_2".to_string()));
-    assert!(names.contains(&"even_4".to_string()));
-}
-
-#[test]
-fn for_loop_zero_iterations_when_from_eq_to() {
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=3, to=3) {
-                box "shouldnt_appear" ()
-            }
-        }"#,
-    );
-    assert!(g.nodes.iter().all(|n| n.name != "shouldnt_appear"));
-}
-
-#[test]
-fn for_loop_step_zero_errors() {
-    let ast = parse(r#"scene { for (var="i", from=0, to=5, step=0) { box "b" () } }"#).expect("parse");
-    let err = lower(&ast).expect_err("step=0 must error");
-    assert!(format!("{err}").contains("step must not be zero"));
-}
-
-#[test]
-fn for_loop_negative_step_iterates_downward() {
-    // Mirrors Python's `range(5, 1, -1)` → 5, 4, 3, 2 (open on the bound).
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=5, to=1, step=-1) {
-                box "down_$i" (size=[0.1, 0.1, 0.1])
-            }
-        }"#,
-    );
-    let names: Vec<_> = g.nodes.iter().filter(|n| n.name.starts_with("down_")).map(|n| n.name.clone()).collect();
-    assert_eq!(names.len(), 4, "from=5, to=1, step=-1 should emit 5,4,3,2");
-    assert!(names.contains(&"down_5".to_string()));
-    assert!(names.contains(&"down_2".to_string()));
-    assert!(!names.contains(&"down_1".to_string()), "to is exclusive");
-}
-
-#[test]
-fn for_loop_iteration_cap_protects_against_runaway_input() {
-    // A `for` loop with a million iterations would grind the host. The
-    // expand-time cap turns it into a friendly error instead.
-    let ast = parse(r#"scene { for (var="i", from=0, to=1000000) { box "b_$i" () } }"#).expect("parse");
-    let err = lower(&ast).expect_err("iteration cap must fire");
-    let msg = format!("{err}");
-    assert!(
-        msg.contains("iteration cap") || msg.contains("cap of"),
-        "expected cap error, got: {msg}"
-    );
-}
-
-#[test]
-fn if_truthy_cond_emits_then_branch() {
-    let g = lower_src(
-        r#"scene {
-            if (cond=1) {
-                box "yes" (size=[1, 1, 1])
-            }
-        }"#,
-    );
-    assert!(g.nodes.iter().any(|n| n.name == "yes"));
-}
-
-#[test]
-fn if_falsy_cond_skips_then_branch() {
-    let g = lower_src(
-        r#"scene {
-            if (cond=0) {
-                box "no" (size=[1, 1, 1])
-            }
-            box "after" (size=[1, 1, 1])
-        }"#,
-    );
-    assert!(g.nodes.iter().all(|n| n.name != "no"));
-    assert!(g.nodes.iter().any(|n| n.name == "after"));
-}
-
-#[test]
-fn if_else_picks_else_branch_when_falsy() {
-    let g = lower_src(
-        r#"scene {
-            if (cond=0) {
-                box "then_branch" ()
-            }
-            else {
-                box "else_branch" ()
-            }
-        }"#,
-    );
-    assert!(g.nodes.iter().all(|n| n.name != "then_branch"));
-    assert!(g.nodes.iter().any(|n| n.name == "else_branch"));
-}
-
-#[test]
-fn else_without_preceding_if_errors() {
-    let ast = parse(r#"scene { else { box "lonely" () } }"#).expect("parse");
-    let err = lower(&ast).expect_err("else without if must error");
-    assert!(format!("{err}").contains("`else` must immediately follow"));
-}
-
-#[test]
-fn comparison_operator_in_cond_works() {
-    // `cond=$n > 1` is the canonical authoring shape for "draw extras only
-    // when there's more than one of something".
-    let make = |n: i32| {
-        let src = format!(
-            r#"module "demo" (n=1) {{
-                box "always" ()
-                if (cond=$n > 1) {{ box "many" () }}
-            }}
-            scene {{ use "demo" (n={n}) }}"#
-        );
-        lower_src(&src)
-    };
-    let one = make(1);
-    let two = make(2);
-    assert!(one.nodes.iter().all(|n| n.name != "many"));
-    assert!(two.nodes.iter().any(|n| n.name == "many"));
-}
-
-#[test]
-fn string_interpolation_in_node_name() {
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=0, to=3) {
-                box "leg_$i" (size=[0.05, 0.5, 0.05])
-            }
-        }"#,
-    );
-    let names: Vec<_> = g.nodes.iter().map(|n| n.name.clone()).collect();
-    assert!(names.contains(&"leg_0".to_string()));
-    assert!(names.contains(&"leg_1".to_string()));
-    assert!(names.contains(&"leg_2".to_string()));
-    // Integer-valued var should NOT render as "leg_0.0".
-    assert!(!names.contains(&"leg_0.0".to_string()));
-}
-
-#[test]
-fn string_interpolation_with_braces() {
-    // ${name} form delimits the binding so authors can compose names like
-    // "${prefix}_panel" without the parser mistaking the underscore for
-    // part of the binding name.
-    let g = lower_src(
-        r#"module "panel" (i=0) {
-            box "${i}_panel" (size=[0.5, 0.1, 0.5])
-        }
-        scene { use "panel" (i=7) }"#,
-    );
-    assert!(g.nodes.iter().any(|n| n.name == "7_panel"));
-}
-
-#[test]
-fn string_interpolation_preserves_multibyte_utf8() {
-    // Regression test for the second-pass review: `interpolate_string`
-    // walked source bytes and used `byte as char`, which corrupts every
-    // UTF-8 continuation byte (0x80..=0xBF) into a U+0080..U+00FF
-    // codepoint. `"pièce_$i"` would render as `"piÃ¨ce_0"` etc. The fix
-    // flushes literal runs via string slicing so multi-byte chars stay
-    // intact.
-    let g = lower_src(
-        r#"scene {
-            for (var="i", from=0, to=2) {
-                box "pièce_$i" (size=[0.05, 0.5, 0.05])
-            }
-        }"#,
-    );
-    let names: Vec<_> = g.nodes.iter().map(|n| n.name.clone()).collect();
-    assert!(
-        names.contains(&"pièce_0".to_string()),
-        "multi-byte char must survive interpolation; got names: {names:?}"
-    );
-    assert!(names.contains(&"pièce_1".to_string()));
-    // Also exercise the `${name}` branch with non-ASCII content around it.
-    let g2 = lower_src(
-        r#"module "p" (i=0) {
-            box "${i}号_部品" (size=[0.5, 0.1, 0.5])
-        }
-        scene { use "p" (i=42) }"#,
-    );
-    assert!(
-        g2.nodes.iter().any(|n| n.name == "42号_部品"),
-        "CJK + ${{name}} interpolation must preserve trailing multi-byte chars; got: {:?}",
-        g2.nodes.iter().map(|n| &n.name).collect::<Vec<_>>()
-    );
-}
-
-#[test]
-fn module_with_for_inside_expands_per_use() {
-    // `for` inside a module body should expand against the call's params.
-    let g = lower_src(
-        r#"module "fence" (posts=1) {
-            for (var="i", from=0, to=$posts) {
-                box "post_$i" (size=[0.05, 0.6, 0.05], pos=[$i * 0.4, 0, 0])
-            }
-        }
-        scene { use "fence" (posts=3) }"#,
-    );
-    let posts: Vec<_> = g.nodes.iter().filter(|n| n.name.starts_with("post_")).collect();
-    assert_eq!(posts.len(), 3);
-}
-
-#[test]
-fn coil_winds_through_full_revolution_via_dsl() {
-    // End-to-end check: parse → validate → lower → mesh. A 1-turn coil
-    // must reach all four cardinal sweeps in XZ; if `coil_mesh` ever
-    // started silently early-returning (turns clamp, samples too low),
-    // this assertion would catch it before the geometry test does.
-    let g = lower_src(
-        r#"scene { coil "spring" (radius=0.5, height=1.0, turns=1, profile_radius=0.05, samples=24) }"#,
-    );
-    let n = find_mesh_node(&g, "spring");
-    let mesh = n.mesh.as_ref().expect("coil produced no mesh");
-    let max_x = mesh.positions.iter().map(|p| p[0]).fold(f32::NEG_INFINITY, f32::max);
-    let min_x = mesh.positions.iter().map(|p| p[0]).fold(f32::INFINITY, f32::min);
-    let max_z = mesh.positions.iter().map(|p| p[2]).fold(f32::NEG_INFINITY, f32::max);
-    let min_z = mesh.positions.iter().map(|p| p[2]).fold(f32::INFINITY, f32::min);
-    assert!(max_x > 0.4 && min_x < -0.4, "coil missing X sweep: [{min_x}, {max_x}]");
-    assert!(max_z > 0.4 && min_z < -0.4, "coil missing Z sweep: [{min_z}, {max_z}]");
-    // Helix climbs from 0 to height — small slack for the parallel-transport
-    // tilt at the endpoints.
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    assert!(min_y.abs() < 0.1 && (max_y - 1.0).abs() < 0.1, "coil Y range wrong: [{min_y}, {max_y}]");
-}
-
-#[test]
-fn coil_handedness_string_validates() {
-    // Both spellings should parse and lower cleanly.
-    let _ = lower_src(
-        r#"scene { coil "lh" (radius=0.3, height=0.6, turns=2, handedness="left") }"#,
-    );
-    let _ = lower_src(
-        r#"scene { coil "rh" (radius=0.3, height=0.6, turns=2, handedness="right") }"#,
-    );
-}
-
-#[test]
-fn coil_unknown_handedness_errors_at_lower() {
-    use crate::parser::parse;
-    let ast = parse(
-        r#"scene { coil "x" (radius=0.3, height=0.6, turns=2, handedness="diagonal") }"#,
-    ).expect("parse");
-    let err = lower(&ast).expect_err("expected lowering error for bad handedness");
-    assert!(format!("{err:#}").contains("handedness"), "wrong error: {err:#}");
-}
-
-#[test]
-fn wave_deformer_displaces_dense_plane() {
-    // A coarse plane has only 4 corner vertices, so to actually see the
-    // wave displacement we use a dense `curved_plane` (zero bend, lots of
-    // segments). The wave attribute should produce non-trivial Y movement
-    // along the X axis on at least some interior vertices.
-    let g = lower_src(
-        r#"scene { curved_plane "water" (size=[4, 1], segments_u=32, segments_v=8,
-            wave=0.15, wave_frequency=0.5, wave_axis="x") }"#,
-    );
-    let mesh = find_mesh_node(&g, "water").mesh.as_ref().unwrap();
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    // A flat plane has y=0 everywhere; a `wave=0.15` has a peak crest of
-    // ~0.15 m. A noticeable spread means the wave actually fired.
-    assert!(
-        max_y - min_y > 0.05,
-        "wave should have produced visible Y spread: [{min_y}, {max_y}]",
-    );
-}
-
-#[test]
-fn wave_axis_string_attribute_lowers() {
-    // `wave_axis="x"` is the canonical valid case — confirms the deformer
-    // attribute is wired through the lowering path. The invalid-string
-    // case (e.g. `wave_axis="diagonal"`) is currently a validator warning
-    // rather than a lowering error: the deformer's `parse_axis` returns
-    // `None` and lowering falls back to X, so a typo builds and ships.
-    // If a future change tightens that to a hard error, add a separate
-    // test asserting `lower_src(...invalid...).is_err()` here.
-    let g = lower_src(
-        r#"scene { curved_plane "ok" (size=[2, 1], segments_u=16, segments_v=4,
-            wave=0.05, wave_axis="x") }"#,
-    );
-    assert!(find_mesh_node(&g, "ok").mesh.is_some());
-}
-
-#[test]
-fn heightfield_lowers_with_displacement() {
-    // The defining property: a heightfield with non-zero amplitude must
-    // produce a Y-spread mesh — proves the noise sampling actually fires
-    // through the lowering path.
-    let g = lower_src(
-        r#"scene { heightfield "ground" (size=[4, 4], segments_u=16, segments_v=16,
-            amplitude=0.5, octaves=3, frequency=0.6, seed=7) }"#,
-    );
-    let mesh = find_mesh_node(&g, "ground").mesh.as_ref().unwrap();
-    let max_y = mesh.positions.iter().map(|p| p[1]).fold(f32::NEG_INFINITY, f32::max);
-    let min_y = mesh.positions.iter().map(|p| p[1]).fold(f32::INFINITY, f32::min);
-    assert!(max_y - min_y > 0.1, "heightfield Y spread too small: [{min_y}, {max_y}]");
-    // 17×17 vertex grid → 289 verts.
-    assert_eq!(mesh.positions.len(), 17 * 17);
-}
-
-#[test]
-fn heightfield_zero_amplitude_is_flat() {
-    let g = lower_src(
-        r#"scene { heightfield "flat" (size=[2, 2], segments_u=4, segments_v=4, amplitude=0) }"#,
-    );
-    let mesh = find_mesh_node(&g, "flat").mesh.as_ref().unwrap();
-    for p in &mesh.positions {
-        assert!(p[1].abs() < 1e-5, "expected flat patch, got y={}", p[1]);
-    }
-}
-
-#[test]
-fn bezier_patch_corners_match_corner_control_points() {
-    // Defining property of bicubic Bézier: P(0,0)=P00, P(1,0)=P30,
-    // P(0,1)=P03, P(1,1)=P33. End-to-end test through the DSL — proves
-    // we wired the `points=` flat list into the right row-major layout.
-    let src = r#"scene {
-        bezier_patch "p" (
-            points = [
-                [-1, 0, -1], [-0.3, 0, -1], [0.3, 0, -1], [1, 0, -1],
-                [-1, 0, -0.3], [-0.3, 1, -0.3], [0.3, 1, -0.3], [1, 0, -0.3],
-                [-1, 0,  0.3], [-0.3, 1,  0.3], [0.3, 1,  0.3], [1, 0,  0.3],
-                [-1, 0,  1], [-0.3, 0,  1], [0.3, 0,  1], [1, 0,  1],
-            ],
-            segments_u = 4, segments_v = 4,
-        )
-    }"#;
-    let g = lower_src(src);
-    let mesh = find_mesh_node(&g, "p").mesh.as_ref().unwrap();
-    // 5×5 vertex grid, row-major along u (5 vertices per row).
-    let nu = 5usize;
-    let nv = 5usize;
-    assert_eq!(mesh.positions.len(), nu * nv);
-    let p00 = mesh.positions[0];
-    let p10 = mesh.positions[(nu - 1) * nv];
-    let p01 = mesh.positions[nv - 1];
-    let p11 = mesh.positions[(nu - 1) * nv + nv - 1];
-    // Corners of the control net are at ±1 on X and Z, y=0.
-    let close = |a: f32, b: f32| (a - b).abs() < 1e-4;
-    assert!(close(p00[0], -1.0) && close(p00[2], -1.0));
-    assert!(close(p10[0], -1.0) && close(p10[2],  1.0));
-    assert!(close(p01[0],  1.0) && close(p01[2], -1.0));
-    assert!(close(p11[0],  1.0) && close(p11[2],  1.0));
-    assert!(p00[1].abs() < 1e-4 && p11[1].abs() < 1e-4);
-    // Centre vertex bulges up — interior control points are at y=1.
-    let centre = mesh.positions[(nu / 2) * nv + nv / 2];
-    assert!(centre[1] > 0.4, "expected centre to bulge, got {}", centre[1]);
-}
-
-#[test]
-fn bezier_patch_wrong_point_count_errors() {
-    use crate::parser::parse;
-    let ast = parse(
-        r#"scene { bezier_patch "x" (points=[[0, 0, 0], [1, 0, 0], [0, 0, 1]]) }"#,
-    ).expect("parse");
-    let err = lower(&ast).expect_err("bezier_patch with 3 points must reject at lower");
-    assert!(format!("{err:#}").contains("16"), "wrong error: {err:#}");
-}
-
-#[test]
-fn metaball_three_centres_unite_into_one_mesh() {
-    // Three overlapping spheres along X with a smooth blend should
-    // produce a single connected mesh whose extent matches the union.
-    let g = lower_src(
-        r#"scene { metaball "blob" (
-            points = [[-0.4, 0, 0], [0, 0, 0], [0.4, 0, 0]],
-            radius = 0.4, blend = 0.1
-        ) }"#,
-    );
-    let mesh = find_mesh_node(&g, "blob").mesh.as_ref().unwrap();
-    assert!(!mesh.positions.is_empty(), "metaball produced no vertices");
-    let max_x = mesh.positions.iter().map(|p| p[0]).fold(f32::NEG_INFINITY, f32::max);
-    let min_x = mesh.positions.iter().map(|p| p[0]).fold(f32::INFINITY, f32::min);
-    assert!(max_x > 0.7, "metaball missing +X extent: {max_x}");
-    assert!(min_x < -0.7, "metaball missing -X extent: {min_x}");
-}
-
-#[test]
-fn metaball_per_point_radii_apply() {
-    let g = lower_src(
-        r#"scene { metaball "asym" (
-            points = [[-1, 0, 0], [1, 0, 0]],
-            radii = [0.3, 0.7], blend = 0
-        ) }"#,
-    );
-    let mesh = find_mesh_node(&g, "asym").mesh.as_ref().unwrap();
-    let max_x = mesh.positions.iter().map(|p| p[0]).fold(f32::NEG_INFINITY, f32::max);
-    let min_x = mesh.positions.iter().map(|p| p[0]).fold(f32::INFINITY, f32::min);
-    assert!((max_x - 1.7).abs() < 0.05, "max_x off: {max_x}");
-    assert!((min_x - (-1.3)).abs() < 0.05, "min_x off: {min_x}");
-}
-
-#[test]
-fn metaball_no_radius_or_radii_errors() {
-    use crate::parser::parse;
-    let ast = parse(
-        r#"scene { metaball "x" (points=[[0, 0, 0]]) }"#,
-    ).expect("parse");
-    let err = lower(&ast).expect_err("metaball without radius/radii must reject");
-    assert!(format!("{err:#}").contains("radius"), "wrong error: {err:#}");
-}
-
-#[test]
-fn vertical_gradient_bakes_endpoint_colours_at_y_extents() {
-    // `vertical` is sugar for `linear(axis=y)`. The bake samples each vertex
-    // against the mesh-local AABB, so the bottom vertices land on `from` and
-    // the top vertices land on `to` regardless of mesh placement.
-    let g = lower_src(
-        r#"
-        material "ramp" (color=[1, 1, 1],
-                         gradient=vertical(from=[1, 0, 0], to=[0, 0, 1]))
-        scene { box "b" (size=[1, 1, 1], pos=[5, 5, 5], mat="ramp") }
-        "#,
-    );
-    let mesh = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    assert_eq!(mesh.colors.len(), mesh.positions.len(), "COLOR_0 row count mismatch");
-    let ymin_i = mesh
-        .positions
-        .iter()
-        .enumerate()
-        .min_by(|a, b| a.1[1].partial_cmp(&b.1[1]).unwrap())
-        .map(|(i, _)| i)
-        .unwrap();
-    let ymax_i = mesh
-        .positions
-        .iter()
-        .enumerate()
-        .max_by(|a, b| a.1[1].partial_cmp(&b.1[1]).unwrap())
-        .map(|(i, _)| i)
-        .unwrap();
-    assert!((mesh.colors[ymin_i][0] - 1.0).abs() < 1e-5);
-    assert!((mesh.colors[ymax_i][2] - 1.0).abs() < 1e-5);
-}
-
-#[test]
-fn radial_gradient_orders_colours_by_distance_from_centre() {
-    // Radial sampling is centre-to-furthest-corner. MoGen primitives don't
-    // generally emit a vertex at their AABB centroid, so a strict t=0 check
-    // would fail. Instead verify the ordering invariant the bake guarantees:
-    // the vertex closest to the AABB centre receives more `center` colour
-    // than the vertex furthest from it.
-    let g = lower_src(
-        r#"
-        material "ring" (color=[1, 1, 1],
-                         gradient=radial(center=[1, 0, 0], edge=[0, 0, 1]))
-        scene { cylinder "c" (radius=0.5, height=2, segments=24, mat="ring") }
-        "#,
-    );
-    let mesh = find_mesh_node(&g, "c").mesh.as_ref().unwrap();
-    let mut centroid = [0.0f32; 3];
-    for p in &mesh.positions {
-        centroid[0] += p[0];
-        centroid[1] += p[1];
-        centroid[2] += p[2];
-    }
-    let n = mesh.positions.len() as f32;
-    centroid = [centroid[0] / n, centroid[1] / n, centroid[2] / n];
-    let dist2 = |p: &[f32; 3]| {
-        let d = [p[0] - centroid[0], p[1] - centroid[1], p[2] - centroid[2]];
-        d[0] * d[0] + d[1] * d[1] + d[2] * d[2]
-    };
-    let nearest = (0..mesh.positions.len())
-        .min_by(|&a, &b| dist2(&mesh.positions[a]).partial_cmp(&dist2(&mesh.positions[b])).unwrap())
-        .unwrap();
-    let furthest = (0..mesh.positions.len())
-        .max_by(|&a, &b| dist2(&mesh.positions[a]).partial_cmp(&dist2(&mesh.positions[b])).unwrap())
-        .unwrap();
-    assert!(
-        mesh.colors[nearest][0] > mesh.colors[furthest][0],
-        "nearest vertex should carry more `center` red"
-    );
-    assert!(
-        mesh.colors[furthest][2] > mesh.colors[nearest][2],
-        "furthest vertex should carry more `edge` blue"
-    );
-}
-
-#[test]
-fn stops_gradient_default_positions_are_evenly_spaced() {
-    // With `colors=[a, b, c]` and no `positions`, stops should land at 0, 0.5, 1.
-    // Verify by giving a tall box a 3-stop Y ramp and checking the mid-Y
-    // vertex receives the middle colour.
-    let g = lower_src(
-        r#"
-        material "rgb" (color=[1, 1, 1],
-                        gradient=stops(colors=[[1, 0, 0], [0, 1, 0], [0, 0, 1]], axis=y))
-        scene { box "b" (size=[1, 4, 1], mat="rgb") }
-        "#,
-    );
-    let mesh = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    // Box verts only exist at the AABB corners (y=±2 for a height-4 box), so
-    // we can't sample at Y=0 directly. Instead verify the endpoints are exact
-    // — those are what default position spacing pins.
-    let ymin_color = mesh
-        .colors
-        .iter()
-        .zip(mesh.positions.iter())
-        .min_by(|a, b| a.1[1].partial_cmp(&b.1[1]).unwrap())
-        .map(|(c, _)| *c)
-        .unwrap();
-    let ymax_color = mesh
-        .colors
-        .iter()
-        .zip(mesh.positions.iter())
-        .max_by(|a, b| a.1[1].partial_cmp(&b.1[1]).unwrap())
-        .map(|(c, _)| *c)
-        .unwrap();
-    assert!((ymin_color[0] - 1.0).abs() < 1e-5, "bottom should be red, got {ymin_color:?}");
-    assert!((ymax_color[2] - 1.0).abs() < 1e-5, "top should be blue, got {ymax_color:?}");
-}
-
-#[test]
-fn material_without_gradient_leaves_mesh_colors_empty() {
-    // Make sure a plain material doesn't accidentally trigger the bake.
-    let g = lower_src(
-        r#"
-        material "plain" (color=[0.5, 0.5, 0.5])
-        scene { box "b" (size=[1, 1, 1], mat="plain") }
-        "#,
-    );
-    let mesh = find_mesh_node(&g, "b").mesh.as_ref().unwrap();
-    assert!(mesh.colors.is_empty(), "plain material should not bake colours");
-}
-
-#[test]
-fn gradient_axis_must_be_x_y_or_z() {
-    use crate::parser::parse;
-    let ast = parse(
-        r#"
-        material "bad" (color=[1, 1, 1],
-                        gradient=linear(from=[1, 0, 0], to=[0, 0, 1], axis=q))
-        scene { box "b" (size=[1, 1, 1], mat="bad") }
-        "#,
-    ).expect("parse");
-    let err = lower(&ast).expect_err("unknown axis must reject");
-    let msg = format!("{err:#}");
-    assert!(msg.contains("axis"), "wrong error: {msg}");
-}
